Remove CheckPayoutStatusJob and update various components
This commit removes the CheckPayoutStatusJob and updates various other components throughout the codebase, including app.js, app/Http/Controllers/PatternController.php, resources/js/Pages/Dashboard/Cart.vue etc. These changes aim to improve program functionality and efficiency. For instance, Placeholders were added for Login fields, 'count' field was removed from orders table, changes were made in the pattern deletion process to prevent the deletion of patterns linked with active orders, 'slug' was added to the channels table for unique identification, adjustments to exception reporting and database transaction logging to improve error handling and debugging.

Took 6 hours 11 minutes

+++ app/Http/Controllers/PatternController.php
@@ -8,6 +8,7 @@ use App\Models\Pattern;
 use App\Models\User;
 use App\Services\AvatarService;
 use App\Services\MediaItemService;
+use Illuminate\Database\QueryException;
 use Illuminate\Http\JsonResponse;
 use Spatie\MediaLibrary\MediaCollections\Exceptions\FileCannotBeAdded;
 use Spatie\MediaLibrary\MediaCollections\Exceptions\FileDoesNotExist;
@@ -157,8 +158,14 @@ class PatternController extends Controller
 
     public function destroy(Pattern $pattern): JsonResponse
     {
-        $pattern->delete();
+        if ($pattern->orders()->exists()){
+            return response()->json([
+                'success' => false,
+                'message' => 'Вы не можете удалить этот шаблон, поскольку на него ссылается заказ.',
+            ], 409);
+        }
 
+        $pattern->delete();
         return response()->json(null, 204);
     }
 }
